Don't expect any specific numbers in random test
Fix random test fails on lua 5.4.

The numbers from math.random on lua 5.4 aren't the same sequence as 5.2
(or whenever I wrote this test). Instead of relying on a specific
sequence, just generate a bunch and check that they're within range.

Don't use testeq for each iteration because all that printing is very
slow:
    testeq(0 <= a, true)
    testeq(a < 1, true)